Projectile (basic) logic.
Restructured update() --> Now follows max range properly, based on calculation using where the projectile was shot at, and it's max range. Calculates it in 360, so down, right, left, up max range.
Added mouse location logic in gamecontroller. The projectile will shoot in the direction user wants it to. We can add more to this in the future, this is just a basic model of how it can look.
Furthermore, fixed the timer tracking updates and movement.